requirements.txt
moviepy
yt-dlp
pillow
piper-tts